+++ Core/ViewModels/StickyNoteProcessor.cs
@@ -16,7 +16,7 @@ namespace Atom.GraphProcessor
         Type IGraphElementProcessor.ModelType => modelType;
 
         /// <summary> 唯一标识 </summary>
-        public int ID => Model.id;
+        public long ID => Model.id;
 
         public InternalVector2Int Position
         {
